+++ test/unit/strats.unit.test.ts
@@ -2,12 +2,23 @@ import assert from "assert";
 import { describe, it } from "mocha";
 
 import Mangrove_v2_0_1 from "../../src/assets/core/v2.0.1/Mangrove.json";
+// v1.0.0
 import AaveKandelSeeder_v1_0_0 from "../../src/assets/strats/v1.0.0/AaveKandelSeeder.json";
 import AavePooledRouter_v1_0_0 from "../../src/assets/strats/v1.0.0/AavePooledRouter.json";
 import KandelLib_v1_0_0 from "../../src/assets/strats/v1.0.0/KandelLib.json";
 import KandelSeeder_v1_0_0 from "../../src/assets/strats/v1.0.0/KandelSeeder.json";
 import MangroveOrderRouter_v1_0_0 from "../../src/assets/strats/v1.0.0/MangroveOrder-Router.json";
 import MangroveOrder_v1_0_0 from "../../src/assets/strats/v1.0.0/MangroveOrder.json";
+// v2.0.0-b1.0
+import AaveKandelSeeder_v2_0_0_b1_0 from "../../src/assets/strats/v2.0.0-b1.0/AaveKandelSeeder.json";
+import AavePooledRouter_v2_0_0_b1_0 from "../../src/assets/strats/v2.0.0-b1.0/AavePooledRouter.json";
+import KandelLib_v2_0_0_b1_0 from "../../src/assets/strats/v2.0.0-b1.0/KandelLib.json";
+import KandelSeeder_v2_0_0_b1_0 from "../../src/assets/strats/v2.0.0-b1.0/KandelSeeder.json";
+import MangroveOrderRouter_v2_0_0_b1_0 from "../../src/assets/strats/v2.0.0-b1.0/MangroveOrder-Router.json";
+import MangroveOrder_v2_0_0_b1_0 from "../../src/assets/strats/v2.0.0-b1.0/MangroveOrder.json";
+import RouterProxyFactory_v2_0_0_b1_0 from "../../src/assets/strats/v2.0.0-b1.0/RouterProxyFactory.json";
+import SimpleAaveLogic_v2_0_0_b1_0 from "../../src/assets/strats/v2.0.0-b1.0/SimpleAaveLogic.json";
+import MangroveAmplifier_v2_0_0_b1_0 from "../../src/assets/strats/v2.0.0-b1.0/MangroveAmplifier.json";
 
 import {
   getAaveKandelSeederVersionDeployments,
@@ -18,6 +29,9 @@ import {
   getAllKandelSeederVersionDeploymentsPerNetwork,
   getAllMangroveOrderRouterVersionDeploymentsPerNetwork,
   getAllMangroveOrderVersionDeploymentsPerNetwork,
+  getAllRouterProxyFactoryVersionDeploymentsPerNetwork,
+  getAllSimpleAaveLogicVersionDeploymentsPerNetwork,
+  getAllMangroveAmplifierVersionDeploymentsPerNetwork,
   getKandelLibVersionDeployments,
   getKandelSeederVersionDeployments,
   getLatestAaveKandelSeederPerNetwork,
@@ -26,9 +40,15 @@ import {
   getLatestKandelSeederPerNetwork,
   getLatestMangroveOrderPerNetwork,
   getLatestMangroveOrderRouterPerNetwork,
+  getLatestRouterProxyFactoryPerNetwork,
+  getLatestSimpleAaveLogicPerNetwork,
   getLatestStratContractsPerNetwork,
+  getLatestMangroveAmplifierPerNetwork,
   getMangroveOrderRouterVersionDeployments,
   getMangroveOrderVersionDeployments,
+  getRouterProxyFactoryVersionDeployments,
+  getSimpleAaveLogicVersionDeployments,
+  getMangroveAmplifierVersionDeployments,
 } from "../../src/strats";
 import { expect } from "chai";
 import { firstVersionDeploymentsToVersionNetworkDeployment } from "./unitTestUtil";
@@ -40,9 +60,9 @@ describe("strats.ts", () => {
         const result = getAaveKandelSeederVersionDeployments({
           released: undefined,
         });
-        assert.equal(result, AaveKandelSeeder_v1_0_0);
+        assert.equal(result, AaveKandelSeeder_v2_0_0_b1_0);
         // NB: Add older old versions here
-        [].forEach((version) => {
+        [AaveKandelSeeder_v1_0_0].forEach((version) => {
           assert.notEqual(result, version);
         });
       });
@@ -55,7 +75,7 @@ describe("strats.ts", () => {
             released: undefined,
           }),
         ).to.deep.equal({
-          "80001": [AaveKandelSeeder_v1_0_0],
+          "80001": [AaveKandelSeeder_v2_0_0_b1_0, AaveKandelSeeder_v1_0_0],
         });
       });
     });
@@ -66,7 +86,7 @@ describe("strats.ts", () => {
           getLatestAaveKandelSeederPerNetwork({ released: undefined }),
         ).to.deep.equal({
           "80001": firstVersionDeploymentsToVersionNetworkDeployment(
-            AaveKandelSeeder_v1_0_0,
+            AaveKandelSeeder_v2_0_0_b1_0,
             "80001",
           ),
         });
@@ -80,9 +100,9 @@ describe("strats.ts", () => {
         const result = getAavePooledRouterVersionDeployments({
           released: undefined,
         });
-        assert.equal(result, AavePooledRouter_v1_0_0);
+        assert.equal(result, AavePooledRouter_v2_0_0_b1_0);
         // NB: Add older old versions here
-        [].forEach((version) => {
+        [AavePooledRouter_v1_0_0].forEach((version) => {
           assert.notEqual(result, version);
         });
       });
@@ -95,7 +115,7 @@ describe("strats.ts", () => {
             released: undefined,
           }),
         ).to.deep.equal({
-          "80001": [AavePooledRouter_v1_0_0],
+          "80001": [AavePooledRouter_v2_0_0_b1_0, AavePooledRouter_v1_0_0],
         });
       });
     });
@@ -106,7 +126,7 @@ describe("strats.ts", () => {
           getLatestAavePooledRouterPerNetwork({ released: undefined }),
         ).to.deep.equal({
           "80001": firstVersionDeploymentsToVersionNetworkDeployment(
-            AavePooledRouter_v1_0_0,
+            AavePooledRouter_v2_0_0_b1_0,
             "80001",
           ),
         });
@@ -118,9 +138,9 @@ describe("strats.ts", () => {
     describe("getKandelLibVersionDeployments", () => {
       it("should find the latest deployment first", () => {
         const result = getKandelLibVersionDeployments({ released: undefined });
-        assert.equal(result, KandelLib_v1_0_0);
+        assert.equal(result, KandelLib_v2_0_0_b1_0);
         // NB: Add older old versions here
-        [].forEach((version) => {
+        [KandelLib_v1_0_0].forEach((version) => {
           assert.notEqual(result, version);
         });
       });
@@ -131,7 +151,7 @@ describe("strats.ts", () => {
         expect(
           getAllKandelLibVersionDeploymentsPerNetwork({ released: undefined }),
         ).to.deep.equal({
-          "80001": [KandelLib_v1_0_0],
+          "80001": [KandelLib_v2_0_0_b1_0, KandelLib_v1_0_0],
         });
       });
     });
@@ -142,7 +162,7 @@ describe("strats.ts", () => {
           getLatestKandelLibPerNetwork({ released: undefined }),
         ).to.deep.equal({
           "80001": firstVersionDeploymentsToVersionNetworkDeployment(
-            KandelLib_v1_0_0,
+            KandelLib_v2_0_0_b1_0,
             "80001",
           ),
         });
@@ -156,9 +176,9 @@ describe("strats.ts", () => {
         const result = getKandelSeederVersionDeployments({
           released: undefined,
         });
-        assert.equal(result, KandelSeeder_v1_0_0);
+        assert.equal(result, KandelSeeder_v2_0_0_b1_0);
         // NB: Add older old versions here
-        [].forEach((version) => {
+        [KandelSeeder_v1_0_0].forEach((version) => {
           assert.notEqual(result, version);
         });
       });
@@ -171,7 +191,7 @@ describe("strats.ts", () => {
             released: undefined,
           }),
         ).to.deep.equal({
-          "80001": [KandelSeeder_v1_0_0],
+          "80001": [KandelSeeder_v2_0_0_b1_0, KandelSeeder_v1_0_0],
         });
       });
     });
@@ -182,7 +202,7 @@ describe("strats.ts", () => {
           getLatestKandelSeederPerNetwork({ released: undefined }),
         ).to.deep.equal({
           "80001": firstVersionDeploymentsToVersionNetworkDeployment(
-            KandelSeeder_v1_0_0,
+            KandelSeeder_v2_0_0_b1_0,
             "80001",
           ),
         });
@@ -196,9 +216,9 @@ describe("strats.ts", () => {
         const result = getMangroveOrderRouterVersionDeployments({
           released: undefined,
         });
-        assert.equal(result, MangroveOrderRouter_v1_0_0);
+        assert.equal(result, MangroveOrderRouter_v2_0_0_b1_0);
         // NB: Add older old versions here
-        [].forEach((version) => {
+        [MangroveOrderRouter_v1_0_0].forEach((version) => {
           assert.notEqual(result, version);
         });
       });
@@ -211,7 +231,10 @@ describe("strats.ts", () => {
             released: undefined,
           }),
         ).to.deep.equal({
-          "80001": [MangroveOrderRouter_v1_0_0],
+          "80001": [
+            MangroveOrderRouter_v2_0_0_b1_0,
+            MangroveOrderRouter_v1_0_0,
+          ],
         });
       });
     });
@@ -222,7 +245,7 @@ describe("strats.ts", () => {
           getLatestMangroveOrderRouterPerNetwork({ released: undefined }),
         ).to.deep.equal({
           "80001": firstVersionDeploymentsToVersionNetworkDeployment(
-            MangroveOrderRouter_v1_0_0,
+            MangroveOrderRouter_v2_0_0_b1_0,
             "80001",
           ),
         });
@@ -236,9 +259,9 @@ describe("strats.ts", () => {
         const result = getMangroveOrderVersionDeployments({
           released: undefined,
         });
-        assert.equal(result, MangroveOrder_v1_0_0);
+        assert.equal(result, MangroveOrder_v2_0_0_b1_0);
         // NB: Add older old versions here
-        [].forEach((version) => {
+        [MangroveOrder_v1_0_0].forEach((version) => {
           assert.notEqual(result, version);
         });
       });
@@ -251,7 +274,7 @@ describe("strats.ts", () => {
             released: undefined,
           }),
         ).to.deep.equal({
-          "80001": [MangroveOrder_v1_0_0],
+          "80001": [MangroveOrder_v2_0_0_b1_0, MangroveOrder_v1_0_0],
         });
       });
     });
@@ -262,7 +285,127 @@ describe("strats.ts", () => {
           getLatestMangroveOrderPerNetwork({ released: undefined }),
         ).to.deep.equal({
           "80001": firstVersionDeploymentsToVersionNetworkDeployment(
-            MangroveOrder_v1_0_0,
+            MangroveOrder_v2_0_0_b1_0,
+            "80001",
+          ),
+        });
+      });
+    });
+  });
+
+  describe("RouterProxyFactory contract", () => {
+    describe("getRouterProxyFactoryVersionDeployments", () => {
+      it("should find the latest deployment first", () => {
+        const result = getRouterProxyFactoryVersionDeployments({
+          released: undefined,
+        });
+        assert.equal(result, RouterProxyFactory_v2_0_0_b1_0);
+        // NB: Add older old versions here
+        [].forEach((version) => {
+          assert.notEqual(result, version);
+        });
+      });
+    });
+
+    describe("getAllRouterProxyFactoryVersionDeploymentsPerNetwork", () => {
+      it("should return all deployments grouped by network", () => {
+        expect(
+          getAllRouterProxyFactoryVersionDeploymentsPerNetwork({
+            released: undefined,
+          }),
+        ).to.deep.equal({
+          "80001": [RouterProxyFactory_v2_0_0_b1_0],
+        });
+      });
+    });
+
+    describe("getLatestRouterProxyFactoryPerNetwork", () => {
+      it("should return the latest deployment for each network", () => {
+        expect(
+          getLatestRouterProxyFactoryPerNetwork({ released: undefined }),
+        ).to.deep.equal({
+          "80001": firstVersionDeploymentsToVersionNetworkDeployment(
+            RouterProxyFactory_v2_0_0_b1_0,
+            "80001",
+          ),
+        });
+      });
+    });
+  });
+
+  describe("SimpleAaveLogic contract", () => {
+    describe("getSimpleAaveLogicVersionDeployments", () => {
+      it("should find the latest deployment first", () => {
+        const result = getSimpleAaveLogicVersionDeployments({
+          released: undefined,
+        });
+        assert.equal(result, SimpleAaveLogic_v2_0_0_b1_0);
+        // NB: Add older old versions here
+        [].forEach((version) => {
+          assert.notEqual(result, version);
+        });
+      });
+    });
+
+    describe("getAllSimpleAaveLogicVersionDeploymentsPerNetwork", () => {
+      it("should return all deployments grouped by network", () => {
+        expect(
+          getAllSimpleAaveLogicVersionDeploymentsPerNetwork({
+            released: undefined,
+          }),
+        ).to.deep.equal({
+          "80001": [SimpleAaveLogic_v2_0_0_b1_0],
+        });
+      });
+    });
+
+    describe("getLatestSimpleAaveLogicPerNetwork", () => {
+      it("should return the latest deployment for each network", () => {
+        expect(
+          getLatestSimpleAaveLogicPerNetwork({ released: undefined }),
+        ).to.deep.equal({
+          "80001": firstVersionDeploymentsToVersionNetworkDeployment(
+            SimpleAaveLogic_v2_0_0_b1_0,
+            "80001",
+          ),
+        });
+      });
+    });
+  });
+
+  describe("MangroveAmplifier contract", () => {
+    describe("getMangroveAmplifierVersionDeployments", () => {
+      it("should find the latest deployment first", () => {
+        const result = getMangroveAmplifierVersionDeployments({
+          released: undefined,
+        });
+        assert.equal(result, MangroveAmplifier_v2_0_0_b1_0);
+        // NB: Add older old versions here
+        [].forEach((version) => {
+          assert.notEqual(result, version);
+        });
+      });
+    });
+
+    describe("getAllMangroveAmplifierVersionDeploymentsPerNetwork", () => {
+      it("should return all deployments grouped by network", () => {
+        expect(
+          getAllMangroveAmplifierVersionDeploymentsPerNetwork({
+            released: undefined,
+          }),
+        ).to.deep.equal({
+          "80001": [MangroveAmplifier_v2_0_0_b1_0],
+        });
+      });
+    });
+
+    describe("getLatestMangroveAmplifierPerNetwork", () => {
+      it("should return the latest deployment for each network", () => {
+        expect(
+          getLatestMangroveAmplifierPerNetwork({ released: undefined }),
+        ).to.deep.equal({
+          "80001": firstVersionDeploymentsToVersionNetworkDeployment(
+            MangroveAmplifier_v2_0_0_b1_0,
             "80001",
           ),
         });
@@ -281,28 +424,40 @@ describe("strats.ts", () => {
             "80001",
           ),
           aaveKandelSeeder: firstVersionDeploymentsToVersionNetworkDeployment(
-            AaveKandelSeeder_v1_0_0,
+            AaveKandelSeeder_v2_0_0_b1_0,
             "80001",
           ),
           aavePooledRouter: firstVersionDeploymentsToVersionNetworkDeployment(
-            AavePooledRouter_v1_0_0,
+            AavePooledRouter_v2_0_0_b1_0,
             "80001",
           ),
           kandelLib: firstVersionDeploymentsToVersionNetworkDeployment(
-            KandelLib_v1_0_0,
+            KandelLib_v2_0_0_b1_0,
             "80001",
           ),
           kandelSeeder: firstVersionDeploymentsToVersionNetworkDeployment(
-            KandelSeeder_v1_0_0,
+            KandelSeeder_v2_0_0_b1_0,
             "80001",
           ),
           mangroveOrderRouter:
             firstVersionDeploymentsToVersionNetworkDeployment(
-              MangroveOrderRouter_v1_0_0,
+              MangroveOrderRouter_v2_0_0_b1_0,
               "80001",
             ),
           mangroveOrder: firstVersionDeploymentsToVersionNetworkDeployment(
-            MangroveOrder_v1_0_0,
+            MangroveOrder_v2_0_0_b1_0,
+            "80001",
+          ),
+          routerProxyFactory: firstVersionDeploymentsToVersionNetworkDeployment(
+            RouterProxyFactory_v2_0_0_b1_0,
+            "80001",
+          ),
+          simpleAaveLogic: firstVersionDeploymentsToVersionNetworkDeployment(
+            SimpleAaveLogic_v2_0_0_b1_0,
+            "80001",
+          ),
+          mangroveAmplifier: firstVersionDeploymentsToVersionNetworkDeployment(
+            MangroveAmplifier_v2_0_0_b1_0,
             "80001",
           ),
         },
@@ -317,6 +472,9 @@ describe("strats.ts", () => {
           kandelSeeder: undefined,
           mangroveOrderRouter: undefined,
           mangroveOrder: undefined,
+          routerProxyFactory: undefined,
+          simpleAaveLogic: undefined,
+          mangroveAmplifier: undefined,
         },
       });
     });